+++ walkers/src/lib.rs
@@ -24,6 +24,9 @@ mod style {
 mod mercator;
 
 #[cfg(feature = "mvt")]
+mod egui_backend;
+
+#[cfg(feature = "mvt")]
 mod expression;
 #[cfg(feature = "mvt")]
 mod mvt;
@@ -37,6 +40,10 @@ pub mod sources;
 mod tiles;
 mod zoom;
 
+#[cfg(feature = "mvt")]
+pub use egui_backend::to_shapes;
+#[cfg(feature = "mvt")]
+pub use egui_backend::wgpu::install_renderer;
 #[cfg(feature = "mvt")]
 pub use expression::Context;
 pub use http_tiles::HttpTiles;
@@ -55,7 +62,12 @@ pub use projector::{
     ScreenProjector,
 };
 #[cfg(feature = "mvt")]
-pub use render::{Geometry, render_line, render_symbol, tessellate_polygon};
+pub use render::drawable::{Drawable, Line, Mesh, Vertex};
+#[cfg(feature = "mvt")]
+pub use render::{
+    Geometry, fill::tessellate_polygon, line::render as render_line,
+    symbol::render as render_symbol,
+};
 pub use style::Style;
 #[cfg(feature = "mvt")]
 pub use style::{